[NSTableView -sizeToFit]: replace floorf with floor
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@11314 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSTableView.m
@@ -2267,7 +2267,7 @@ byExtendingSelection: (BOOL)flag
 		}
 	      else
 		{
-		  int remainingInt = floorf(remainingWidth);
+		  int remainingInt = floor(remainingWidth);
 		  int quotient = remainingInt / numberOfCurrentColumns;
 		  int remainder = remainingInt - quotient * numberOfCurrentColumns;
 		  int oldRemainder = remainder;
